+++ CHANGELOG.md
@@ -15,3 +15,7 @@
 - The compiler now correctly tracks the minimum required version for constant
   record updates to be `>= 1.14.0`.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
+- The compiler now correctly tracks the minimum required version for expressions
+  in `BitArray`s' `size` option to be `>= 1.12.0`.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))